devcor3/m5/requirements.txt
python3-logstash
xmltodict
ncclient
